# Speed-up LSP completion response times