@apiology 2025-08-21
I'd suggest creating a spec demonstrating creating an alias in a subclass to a method defined in the superclass - that'd show that if you can't find the method the easy way, you fall back to the old way