@apiology 2025-08-21
> /home/runner/work/solargraph/solargraph/lib/solargraph/api_map/store.rb:55 - Not enough arguments to Kernel#select

Probably another good spec to add - these select methods are on arrays, which implement Enumerable.  Enumerable#select is an alias for Enumerable#find_all, and I'm guessing the logic to find that alias and use it isn't working right with this change.