@castwide 2025-08-24
Nice work, thanks! Out of curiosity, I tested this branch against a project that had slow completions due to heavy use of dynamic yield receivers. Performance was noticeably improved there, too.

The linting error should go away with the merge. If it doesn't, I'll deal with the fallout.