@castwide 2025-08-25
From what I've determined, there's nothing to be gained from making `Library#sync_catalog` asynchronous. Ruby threads are always executed in a single process, so it will never run concurrently at the OS level (e.g., POSIX). Any other function within the solargraph process that needs the newest maps will need to wait for the library to synchronize anyway. Ruby threads don't make it any faster. The main reason `#sync_catalog` exists is to ensure that clients have access to the most current versions of code maps at the moment they're requested.